+++ nilearn/_utils/class_inspect.py
@@ -1,8 +1,5 @@
 """Small utilities to inspect classes."""
 
-from dataclasses import dataclass
-
-from packaging.version import parse
 from sklearn import __version__ as sklearn_version
 from sklearn.utils.estimator_checks import (
     check_estimator as sklearn_check_estimator,
@@ -30,7 +27,11 @@ else:
     VALID_CHECKS.append("check_estimator_get_tags_default_keys")
 
 
-CHECKS_TO_SKIP_IF_NIIMG_INPUT = {
+# TODO / TOCHECK: with sklearn >= 1.6
+# some of those checks should be skipped 'automatically'
+# by sklearn
+# and could be removed from this list.
+CHECKS_TO_SKIP_IF_IMG_INPUT = {
     "check_estimator_sparse_array",
     "check_estimator_sparse_data",
     "check_estimator_sparse_matrix",
@@ -113,7 +114,7 @@ def check_estimator(estimator=None, valid=True, extra_valid_checks=None):
 
             if (
                 niimg_input or surf_img
-            ) and check.func.__name__ in CHECKS_TO_SKIP_IF_NIIMG_INPUT:
+            ) and check.func.__name__ in CHECKS_TO_SKIP_IF_IMG_INPUT:
                 continue
 
             if valid and check.func.__name__ in valid_checks:
@@ -161,55 +162,3 @@ def get_params(cls, instance, ignore=None):
             params[param_name] = getattr(instance, param_name)
 
     return params
-
-
-ver = parse(sklearn_version)
-if ver.release[1] >= 6:
-    from sklearn.utils import InputTags as SkInputTags
-
-    @dataclass
-    class InputTags(SkInputTags):
-        """Tags for the input data.
-
-        Nilearn version of sklearn.utils.InputTags
-        https://scikit-learn.org/1.6/modules/generated/sklearn.utils.InputTags.html#sklearn.utils.InputTags
-        """
-
-        # same as base input tags of
-        # sklearn.utils.InputTags
-        one_d_array: bool = False
-        two_d_array: bool = True
-        three_d_array: bool = False
-        sparse: bool = False
-        categorical: bool = False
-        string: bool = False
-        dict: bool = False
-        positive_only: bool = False
-        allow_nan: bool = False
-        pairwise: bool = False
-
-        # nilearn specific things
-        # estimator accepts for str, Path to .nii[.gz] file
-        # or NiftiImage object
-        niimg_like: bool = True
-        # estimator accepts SurfaceImage object
-        surf_img: bool = False
-else:
-
-    def tags(niimg_like=True, surf_img=False, **kwargs):
-        """Add nilearn tags to estimator.
-
-        See also: InputTags
-
-        TODO remove when dropping sklearn 1.5
-        """
-        X_types = []
-        if "X_types" in kwargs:
-            X_types = kwargs["X_types"]
-        X_types.append("2darray")
-        if niimg_like:
-            X_types.append("niimg_like")
-        if surf_img:
-            X_types.append("surf_img")
-        X_types = list(set(X_types))
-        return dict(X_types=X_types, **kwargs)
