+++ nilearn/decoding/searchlight.py
@@ -339,11 +339,11 @@ class SearchLight(TransformerMixin, BaseEstimator):
         # bumping sklearn_version > 1.5
         ver = parse(sklearn_version)
         if ver.release[1] < 6:
-            from nilearn._utils.class_inspect import tags
+            from nilearn._utils.tags import tags
 
             return tags()
 
-        from nilearn._utils.class_inspect import InputTags
+        from nilearn._utils.tags import InputTags
 
         tags = super().__sklearn_tags__()
         tags.input_tags = InputTags()
