+++ nilearn/maskers/surface_masker.py
@@ -154,11 +154,11 @@ class SurfaceMasker(TransformerMixin, CacheMixin, BaseEstimator):
         # bumping sklearn_version > 1.5
         ver = parse(sklearn_version)
         if ver.release[1] < 6:
-            from nilearn._utils.class_inspect import tags
+            from nilearn._utils.tags import tags
 
             return tags(surf_img=True, niimg_like=False)
 
-        from nilearn._utils.class_inspect import InputTags
+        from nilearn._utils.tags import InputTags
 
         tags = super().__sklearn_tags__()
         tags.input_tags = InputTags(surf_img=True, niimg_like=False)
